feat: initialize cilium CNI
Lays the groundwork for installing infrastructure into the Kind cluster.
Uses helmfile to declare the helm configuration, then renders it
in-place, so that we can eventually run static analysis tools in CI.

Additionally, it allows us to determine what is going to be installed in
the cluster before actually applying it. This makes for a better GitOps
pattern, and more useful peer reviews on a normal team.

Creates a script to automatically find and template all helmfiles in the
repo. Also installs Cilium via OpenTofu/Terraform as it's needed before
any pod can be scheduled or access network resources.