feat: install tailscale k8s operator
Provide private network access without configuring additional VPNs,
firewall rules, or open ports, regardless of where this is installed,
thanks to tailscale and NAT hole-punching.

Installs tailscale via TF so that we can provision the OAuth client
needed by the operator. ArgoCD will also manage this once bootstrapped.

ArgoCD is additionally exposed over the tailnet at
https://argocd.<tailnet>.ts.net. In my case, the tailnet is tail661100.

I could go back & forth on exposing this both publicly and over the
tailnet. If maximum security is required, and users are aware of the
tailnet FQDN, I would not publicly expose ArgoCD due to the control it
has over the cluster. If some risk of vulnerability in the network stack
of ArgoCD is acceptable, or users are unfamiliar with tailscale, I would
expose it publicly using a good SSO provider and 2FA, thanks to Dex.

In this case, I'm exposing ArgoCD over both to prove networking is
configured correctly over both Tailscale and Cilium/Envoy (via Gateway
API).